requirements.txt
asgiref==3.7.2
blessed==1.17.12
Django==5.0.2
django-ckeditor==6.7.1
django-filter==23.5
django-js-asset==2.2.0
django-tables2==2.7.0
et-xmlfile==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
objexplore==1.6.3
openpyxl==3.1.2
pillow==10.2.0
Pygments==2.17.2
rich==13.7.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
wcwidth==0.2.13